+++ src/entities/index.tsx
@@ -1,0 +1,10 @@
+import Articles from './Articles';
+import LegalServices from './LegalServices';
+import NeedConsultation from './NeedConsultation';
+import PracticalArea from './PracticalArea';
+import QuoteSlider from './QuoteSlider';
+import RequestConsultation from './RequestConsultation';
+import Content from './shared/Content';
+//TODO: import Contacts from './Contacts';
+export { Articles, Content, LegalServices, NeedConsultation, PracticalArea, QuoteSlider, RequestConsultation };
+