+++ src/widgets/index.tsx
@@ -1,0 +1,13 @@
+import Button from './Button';
+import Card from './Card';
+import CenterInfoLine from './CenterInfoLine';
+import Menu from './Menu';
+import MenuItem from './MenuItem';
+import Paragraph from './Paragraph';
+import PhoneNumber from './PhoneNumber';
+import QuoteSlideItem from './QuoteSlideItem';
+import Title from './Title';
+
+
+export { Button, Card, CenterInfoLine, Menu, MenuItem, Paragraph, PhoneNumber, QuoteSlideItem, Title };
+