+++ CHANGELOG.md
@@ -1,9 +1,13 @@
-#### 0.0.4 (2024-11-30)
+#### 0.0.4 (2024-12-01)
 
 ##### New Features
 
-- **NeedConsultation:** add NeedConsultation block (3279d92d)
+- **NeedConsultation:** add NeedConsultation block (f244b28d)
 - **QuoteSlider:** impl + demo (2009f53e)
+
+##### Refactors
+
+- move to fsd (ba8b9667)
 
 #### 0.0.3 (2024-11-30)